+++ spec/spec_helper.rb
@@ -6,6 +6,10 @@ require "timecop"
 require "tmpdir"
 require_relative "support/backspin_helper"
 
+Backspin.configure do |config|
+  config.logger = nil
+end
+
 RSpec.configure do |config|
   config.filter_run_when_matching :focus
   config.disable_monkey_patching!